requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
psycopg2-binary
pydantic
fastapi.security.api_key
requests
pymupdf
